requirements.txt
#cairocffi @ http://www.lfd.uci.edu/~gohlke/pythonlibs/#cairocffi
#maybe we need cairo cffi dependency for linux too
python-pptx
CairoSVG # and pip install pipwin && pipwin install cairocffi
html2text
typing-extensions

#--trusted-host www.lfd.uci.edu --extra-index-url http://www.lfd.uci.edu/~gohlke/pythonlibs/#cairocffi
#cairocffi 